requirements.txt

pytesseract
pillow
opencv-python
pdf2image
numpy
pandas
rapidfuzz